Cleaned up release folder somewhat Fixed (?) the poping up window in CreateProcess
Fixed bug in change of file extension function when given fname like .\\fName

Commented out 
//  madExcept,
//  madLinkDisAsm,
//  madListHardware,
//  madListProcesses,
//  madListModules,

in Delphi project

+++ trunk/releases/xe/include/rr/rrModelFromC.h
@@ -80,7 +80,7 @@ class RR_DECLSPEC ModelFromC : public rrObject
 //        vector<double>                        lp ;            //Local parameters
 
         double*                                 c;                //Compartment volumes
-        double*                                 cSize;                //Compartment volumes
+        int*									cSize;                //Compartment volumes
 
         double*                                 rates;
         int*                                    ratesSize;
